Reorganize the routing logic
The router now uses two reactive state nodes -- essentially each is
a channel with a read end and a write end, and only the read end of
one and the write end of the other are exposed to other components.
This makes the routing logic more comprehensible. Unfortunately
this means that other components when navigating to a new route
must state all the properties of the route at once, not just the
one they want to change, as is usually the case with state in
Solid. This is because there are actually two state updates that go
on. Perhaps must simplify further.